Dev (#14)
* added ide files to gitignore

* Use BLOG_FONT_BODY env var

* package updates, cleanup

* Merge pull request #2 from vulcan-byte/codex/update-readme-to-reference-theme-utils.js

Fix README theme utils reference

* removed leftover code

* small typo

* docs: clarify comment for next post (#3)

* Sanitize theme env vars (#5)

* chore: resolve npm vulnerabilities (#6)

* cleanup

* Update README for App Router (#8)

* Migrate to Next.js app router (#7)

* refactor: migrate to app router

* Remove next/head usage

* router update

* docs: update paths for app router

* needed deprecated field to preview feat

---------


* Add volcano theme (#9)

* Add volcano color theme

* using yarn

---------


* Migrate components to TS (#12)

* refactor: use arrow functions

* work on TS migration

---------


* Convert remaining files to TypeScript (#13)

* refactor: use arrow functions

* work on TS migration

* cleanup

* Update README paths for TypeScript files

---------


* cleanup

* missing dependency

* yarn.lock

* cleanup

---------
