+++ README.md
@@ -80,12 +80,12 @@ https://user-images.githubusercontent.com/3611928/153997545-6dcdeef0-e570-49e7-9
 
 [alt: video walkthrough of editing env vars]
 
-If setting an environment variable isn't your cup of tea, the defaults can be changed in [`src/utils/global-data.js`](/src/utils/global-data.js). You can also remove the variables and hard code blog information where these variables are used in the code base.
+If setting an environment variable isn't your cup of tea, the defaults can be changed in [`src/utils/globalData.ts`](/src/utils/globalData.ts). You can also remove the variables and hard code blog information where these variables are used in the code base.
 
-- `BLOG_THEME, BLOG_FONT_HEADINGS, & BLOG_FONT_BODY` are used in [`src/utils/theme-utils.js`](src/utils/theme-utils.js)
-- `BLOG_NAME, BLOG_TITLE, BLOG_FOOTER_TEXT` are used in [`src/app/page.js`](src/app/page.js) & [`src/app/posts/[slug]/page.js`](src/app/posts/[slug]/page.js) through the `globalData` object.
+- `BLOG_THEME, BLOG_FONT_HEADINGS, & BLOG_FONT_BODY` are used in [`src/utils/themeUtils.ts`](src/utils/themeUtils.ts)
+- `BLOG_NAME, BLOG_TITLE, BLOG_FOOTER_TEXT` are used in [`src/app/page.tsx`](src/app/page.tsx) & [`src/app/posts/[slug]/page.tsx`](src/app/posts/[slug]/page.tsx) through the `globalData` object.
 
-Values provided for `BLOG_THEME`, `BLOG_FONT_HEADINGS`, and `BLOG_FONT_BODY` are sanitized to allow only alphanumeric characters and hyphens. Make sure the values match one of the themes or fonts defined in [`themes.js`](themes.js) to ensure they are applied.
+Values provided for `BLOG_THEME`, `BLOG_FONT_HEADINGS`, and `BLOG_FONT_BODY` are sanitized to allow only alphanumeric characters and hyphens. Make sure the values match one of the themes or fonts defined in [`themes.ts`](themes.ts) to ensure they are applied.
 
 ## Adding new posts
