+++ src/app/posts/[slug]/page.tsx
@@ -1,50 +1,51 @@
-import {getGlobalData} from "../../../utils/global-data";
-import {getNextPostBySlug, getPostBySlug, getPostFilePaths, getPreviousPostBySlug} from "../../../utils/mdx-utils";
-import CustomLink from "../../../components/CustomLink";
-import CustomImage from "../../../components/CustomImage";
+
 import {MDXRemote} from 'next-mdx-remote/rsc';
 import Link from "next/link";
-import ArrowIcon from "../../../components/ArrowIcon";
-import Footer from "../../../components/Footer";
-import Layout, {GradientBackground} from "../../../components/Layout";
-import Header from "../../../components/Header";
+import Footer from "@/components/Footer";
+import Layout, {GradientBackground} from "@/components/Layout";
 import rehypePrettyCode from "rehype-pretty-code";
 import rehypeUnwrapImages from "rehype-unwrap-images";
 import remarkGfm from "remark-gfm";
+import {getGlobalData} from "@/utils/globalData";
+import {getNextPostBySlug, getPostBySlug, getPostFilePaths, getPreviousPostBySlug} from "@/utils/mdxUtils";
+import CustomLink from "@/components/CustomLink";
+import CustomImage from "@/components/CustomImage";
+import Header from "@/components/Header";
+import ArrowIcon from "@/components/ArrowIcon";
 
 
-export async function generateMetadata({params}) {
-  const {slug} = await params;
+export const generateMetadata = async ({ params }) => {
+  const { slug } = await params;
   const globalData = getGlobalData();
-  const {mdxSource, data: frontMatter} = await getPostBySlug(slug);
+  const { data: frontMatter } = await getPostBySlug(slug);
   return {
     title: `${frontMatter.title} - ${globalData.name}`,
     description: frontMatter.description,
   };
-}
+};
 
 // Custom components/renderers to pass to MDX.
-// Since the MDX files aren't loaded by webpack, they have no knowledge of how
+// Since webpack doesn't load the MDX files, they have no knowledge of how
 // to handle import statements. Instead, you must include components in scope
 // here.
 const components = {
   a: CustomLink,
-  // It also works with dynamically-imported components, which is especially
+  // It also works with dynamically imported components, which is especially
   // useful for conditionally loading components for certain routes.
   // See the notes in README.md for more details.
   img: CustomImage,
 };
 
-export default async function PostPage({params}) {
-  const {slug} = await params;
+const PostPage = async ({ params }) => {
+  const { slug } = await params;
   const globalData = getGlobalData();
-  const {content, data} = await getPostBySlug(slug);
+  const { content, data } = await getPostBySlug(slug);
   const prevPost = getPreviousPostBySlug(slug);
   const nextPost = getNextPostBySlug(slug);
   return (
     <Layout>
       <Header name={globalData.name}/>
-      <article className="px-6 md:px-0" data-sb-object-id={`posts/${slug}.mdx`}>
+      <article className="px-6 md:px-0" data-sb-object-id={`data/posts/${slug}.mdx`}>
         <header>
           <h1
             className="mb-12 text-3xl text-center md:text-5xl dark:text-white"
@@ -81,7 +82,7 @@ export default async function PostPage({params}) {
         <div className="grid mt-12 md:grid-cols-2 lg:-mx-24">
           {prevPost && (
             <Link
-              href={`/posts/${prevPost.slug}`}
+              href={`/data/posts/${prevPost.slug}`}
               className="flex flex-col px-10 py-8 text-center transition border border-gray-800/10 bg-white/10 md:text-right first:rounded-t-lg md:first:rounded-tr-none md:first:rounded-l-lg last:rounded-r-lg last:rounded-b-lg backdrop-blur-lg dark:bg-black/30 hover:bg-white/20 dark:hover:bg-black/50 dark:border-white/10 last:border-t md:border-r-0 md:last:border-r md:last:rounded-r-none"
             >
               <p className="mb-4 text-gray-500 uppercase dark:text-white dark:opacity-60">
@@ -95,7 +96,7 @@ export default async function PostPage({params}) {
           )}
           {nextPost && (
             <Link
-              href={`/posts/${nextPost.slug}`}
+              href={`/data/posts/${nextPost.slug}`}
               className="flex flex-col px-10 py-8 text-center transition border border-t-0 border-b-0 border-gray-800/10 bg-white/10 md:text-left md:first:rounded-t-lg last:rounded-b-lg first:rounded-l-lg md:last:rounded-bl-none md:last:rounded-r-lg backdrop-blur-lg dark:bg-black/30 hover:bg-white/20 dark:hover:bg-black/50 dark:border-white/10 first:border-t first:rounded-t-lg md:border-t last:border-b"
             >
               <p className="mb-4 text-gray-500 uppercase dark:text-white dark:opacity-60">
@@ -120,10 +121,11 @@ export default async function PostPage({params}) {
       />
     </Layout>
   );
-}
+};
 
-export async function generateStaticParams() {
-  return getPostFilePaths()
+export default PostPage;
+
+export const generateStaticParams = async () =>
+  getPostFilePaths()
     .map((path) => path.replace(/\.mdx?$/, ''))
-    .map((slug) => ({slug}));
-}
+    .map((slug) => ({ slug }));
