--- src/components/CustomLink.js
@@ -1,5 +1,0 @@
-import Link from 'next/link';
-
-export default function CustomLink({ as, href, ...otherProps }) {
-  return <Link as={as} href={href} {...otherProps}></Link>;
-}
