+++ themes.ts
@@ -1,4 +1,4 @@
-exports.COLOR_THEMES = {
+export const COLOR_THEMES = {
   default: {
     colors: {
       primary: '#7D7AFF',
@@ -55,7 +55,7 @@ exports.COLOR_THEMES = {
   },
 };
 
-exports.FONT_THEMES = {
+export const FONT_THEMES = {
   'sans-serif': `ui-sans-serif, system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, "Helvetica Neue", Arial, "Noto Sans", sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol", "Noto Color Emoji"`,
   serif: `ui-serif, Georgia, Cambria, "Times New Roman", Times, serif`,
   monospace: `ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, "Liberation Mono", "Courier New", monospace`,
